requirements.txt
pandas
flask
pillow
numpy
matplotlib